requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
flake8==7.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
mccabe==0.7.0
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pyasn1==0.4.8
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
win32_setctime==1.2.0
streamlit>=1.25.0
requests>=2.31.0
